Added Edit functionality/Fixed delete
Added the ability to edit entries via the "Edit" and "Save Edit" buttons. Also discovered an issue with the delete button that plagued the edit button. In both cases, it seems that not choosing the Id cell would result in an out of bounds error. Fixed this by changing the DataGridView to select the entire row when selecting a cell. There was never any need for the user to select an individual cell anyway, so this chance is entirely positive